@@ -1,30 +1,32 @@
 ﻿using System;
-using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 using Serilog;
 
 namespace TestAuthorityCore
 {
     /// <summary>
-    /// The one and only.
+    ///     The one and only.
     /// </summary>
     public class Program
     {
         /// <summary>
-        /// The mainest of them all.
+        ///     The mainest of them all.
         /// </summary>
         /// <param name="args">Arguments.</param>
         public static void Main(string[] args)
         {
             Log.Logger = new LoggerConfiguration()
-            .Enrich.FromLogContext()
-            .WriteTo.Console()
-            .CreateLogger();
+                .Enrich.FromLogContext()
+                .WriteTo.Console()
+                .CreateLogger();
 
             try
             {
-                CreateWebHostBuilder(args)
-                    .Build().Run();
+                CreateHostBuilder(args)
+                    .Build()
+                    .Run();
             }
             catch (Exception ex)
             {
@@ -37,13 +39,29 @@ namespace TestAuthorityCore
         }
 
         /// <summary>
-        /// Create a <seecref name="IWebHostBuilder"/>.
+        ///     Create a <seecref name="IWebHostBuilder" />.
         /// </summary>
         /// <param name="args">Arguments</param>
-        /// <returns><seecref name="IWebHostBuilder"/></returns>
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseSerilog()
-                .UseStartup<Startup>();
+        /// <returns>
+        ///     <seecref name="IWebHostBuilder" />
+        /// </returns>
+        public static IHostBuilder CreateHostBuilder(string[] args)
+        {
+            return Host.CreateDefaultBuilder(args)
+                .UseSerilog((context, services, configuration) =>
+                {
+                    configuration
+                        .ReadFrom.Configuration(context.Configuration)
+                        .ReadFrom.Services(services)
+                        .Enrich.FromLogContext()
+                        .WriteTo.Console();
+                })
+                .ConfigureWebHostDefaults(x => x.UseStartup<Startup>())
+                .ConfigureAppConfiguration(((context, builder) =>
+                {
+                    builder.AddJsonFile("appsettings.json", optional: true, reloadOnChange: false);
+                    builder.AddJsonFile($"appsettings.{context.HostingEnvironment.EnvironmentName}.json", optional: true, reloadOnChange: false);
+                }));
+        }
     }
 }
