@@ -1,6 +1,6 @@
 using FluentValidation;
-using TestAuthorityCore.Contracts;
-using TestAuthorityCore.X509;
+using TestAuthority.Domain.Models;
+using CertificateRequestModel = TestAuthorityCore.Contracts.CertificateRequestModel;
 
 namespace TestAuthorityCore.Validators
 {
@@ -14,9 +14,12 @@ namespace TestAuthorityCore.Validators
         /// </summary>
         public CertificateRequestValidator()
         {
-            RuleFor(x => x).Must(AnyHostnamesOrIpAddresses)
-                .WithMessage("You must provide at least one hostname or IP address.");
-            RuleFor(x => x.Password).MinimumLength(1).When(x => x.Format == CertificateFormat.Pfx)
+            RuleFor(x => x.CommonName)
+                .NotEmpty().WithMessage("You must provide a Common Name");
+            RuleFor(x => x)
+                .Must(AnyHostnamesOrIpAddresses).WithMessage("You must provide at least one hostname or IP address.");
+            RuleFor(x => x.Password)
+                .MinimumLength(1).When(x => x.Format == OutputFormat.Pfx)
                 .WithMessage("You must provide a password for PFX");
         }
 
